[MI-3355]: Added dropdown to choose teams and channels (#14)
* [MI-3355]: Removed extra types

* [MI-3355]: Improved message of view action modal

* [MI-3355]: removed extra packages

* [MI-3355]: Added eod to files

* [MI-3355]: Fixed adding action functionality

* [MI-3355]: Fixed the jump in UI during validation warnings

* [MI-3355]: fixed maximum nested callback in useEffect

* [MI-3355]: Fixed validation warnings in Add action modal

* [MI-3355]: Added ellipses to the tables

* [MI-3355]: updated API to fetch data in batches

* [MI-3355]: Removed comment and fixed empty lines

* [MI-3355]: Fixed css of modal visibility

* [MI-3355]: Added svg icons to different file

* [MI-3355]: Improvised the folder structure

* [MI-3355]: Improved the variable name in existingConfigTable

* [MI-3355]: Fixed error in adding new action to config with index zero

* [MI-3355]: Fixed import issue of .d.ts files

* [MI-3355]: Used fat-arrow function instead of normal functions

* [MI-3355]: Installed types and packages for js-cookie

* [MI-3355]: Setup reduxstore

* [MI-3355]: Disabled dropdown during loading state

* [MI-3355]: Improved getBaseUrl method

* [MI-3355]: Improved the API call

* [MI-3355]: Added constants and used baseUrl from redux state

* [MI-3355]: Fixed jump in UI in edit action modal

* [MI-3355]: Joined array of string with comma for messages & disabled dropdown incase of API failure

* [MI-3355]: Added type of onChange function

* [MI-3355]: Removed extra variables

* [MI-3355]: Improved the validation conditions

* [MI-3355]: Fixed css according to reviews by abhishek

* [MI-3355]: Improved visibility of action successfull message

* [MI-3355]: Used scss instead of css

* [MI-3355]: Fixed the CSS of channel-added-to content

* [MI-3355]: Added semicolons to types in common.d.ts

* Update webapp/src/components/modals/actionModal.tsx


* Update webapp/src/components/tables/existingConfigTable.tsx


* Update webapp/src/components/svgIcons/svg.tsx


* [MI-3355]: Used px instead of % in css

* [MI-3355]: Fixed some minor css

* [MI-3355]: Removed extra spaces from existingConfigTable

---------
