+++ webapp/package.json
@@ -61,12 +61,15 @@
     "webpack-cli": "3.3.12"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "1.9.5",
     "core-js": "3.6.5",
+    "js-cookie": "3.0.5",
     "mattermost-redux": "5.27.0",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
     "react-bootstrap": "2.8.0",
+    "react-dom": "18.2.0",
     "react-redux": "7.2.0",
+    "react-select": "5.7.4",
     "redux": "4.0.5",
     "typescript": "5.1.6"
   },
