+++ webapp/src/constants/apiConstants.ts
@@ -1,0 +1,7 @@
+export const getChannel = '/channels?exclude_default_channels=true';
+export const getTeam = '/teams';
+export const apiVersion1 = '/api/v1';
+export const apiVersion4 = '/api/v4';
+export const plugins = '/plugins';
+export const mmAuthToken = 'MMAUTHTOKEN';
+export const bearer = 'Bearer ';
