--- webapp/src/containers/components/tables/styles.css
@@ -1,118 +1,0 @@
-.existing-config-table.btn-group {
-    display: flex;
-    justify-content: space-between;
-    gap: 10px;
-}
-
-.existing-config-table.table.table tbody tr td,
-.existing-config-table.table.table thead tr th,
-.existing-config-table.table.table thead {
-  border: 2px solid #dddddd;
-  vertical-align: middle;
-}
-  
-.existing-config-table.table td {
-    text-align: center;
-    vertical-align: middle;
-}
-
-.existing-config-table.table th {
-    text-align: center;
-    vertical-align: bottom;
-} 
-
-.existing-config-table.table {
-    max-width: 100%;
-}
-
-.existing-config-table.table thead tr th td{
-    white-space:nowrap
-}
-
-.existing-config-table.btn-group {
-    gap: 1px;
-}
-
-.existing-config-table.btn-group .btn.btn-default {
-    background: none;
-    border-color: transparent;
-    border: transparent;
-}
-
-.existing-config-table .btn-group.btn.btn-default:hover, .btn-group .btn.btn-default:focus {
-    background: transparent;
-    color: #000000;
-}
-
-.add-config-btn {
-    margin: 1px 0 0 0;
-}
-
-.config {
-    padding-left: 15px;
-}
-
-.name {
-    font-weight: 600;
-    font-size: 14px;
-    margin-bottom: 1em;
-}
-
-.team-name {
-    max-width: 15%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-}
-
-.delay {
-    width: 10%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-}
-
-.message {
-    width: 55%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-}
-
-.include-guests {
-    width: 10%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-}
-
-.option {
-    width: 10%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-}
-
-.guest-value {
-    text-align: center;
-    vertical-align: middle;
-}
-
-.delay-value {
-    text-align: center;
-    vertical-align: middle;
-}
-
-.team-name-value {
-    text-align: center;
-    vertical-align: middle;
-}
-
-.existing-config-table.table .messageValue {
-    text-align: left;
-    vertical-align: middle;
-}
-
-.message-value .td {
-    text-align: left;
-}
