+++ webapp/src/index.tsx
@@ -1,11 +1,9 @@
 import {Store, Action} from 'redux';
-
 import {GlobalState} from 'mattermost-redux/types/store';
 
-// eslint-disable-next-line import/no-unresolved
 import {PluginRegistry} from 'types/mattermostWebapp';
 
-import ExistingConfigTable from 'containers/components/tables/existingConfigTable';
+import ExistingConfigTable from 'components/tables/existingConfigTable';
 
 import {id} from './manifest';
