+++ webapp/src/types/plugin/common.d.ts
@@ -4,7 +4,7 @@ type Actions = {
     actionDisplayName: string,
     channelsAddedTo: string[],
     actionSuccessfullMessage: string[]
-}
+};
 
 type Configs = {
     teamName: string,
@@ -13,4 +13,24 @@ type Configs = {
     includeGuests: string | null,
     attachmentMessage: string[] | null,
     actions: Actions[] | null
-}
+};
+
+type GroupTypes = {
+    label: string;
+    value: string;
+};
+
+type OptionTypes = {
+    value: string;
+    label: string;
+    data: string;
+};
+
+type Teams = {
+    display_name: string;
+};
+
+type Channels = {
+    display_name: string;
+    team_name: string,
+};
